Prefer locally-installed packages over packages available in the repos
This is because without this preference, if there's metadata changes
between the currently-installed version and the version from the repo
(i.e. different dependency list), apt may decide to install the version
from the repo if it determines that that is better/easier to install any
dependencies. This preference file says that the local version is better
than the repo version, so this package is "better".

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>